Split scss files
