+++ package.json
@@ -7,7 +7,9 @@
   "license": "GPL-3.0",
   "dependencies": {
     "@cypress/react": "^5.12.4",
+    "@types/classnames": "^2.3.1",
     "bulma": "^0.9.3",
+    "classnames": "^2.3.1",
     "postcss": "^8.4.12",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
