+++ src/components/List/List.scss
@@ -1,0 +1,35 @@
+@import "./../../styles/fonts.scss";
+@import "./../../styles/variables.scss";
+
+.goods {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+
+  &__item {
+    font-size: 20px;
+    user-select: none;
+  }
+
+  &__buttons {
+    display: flex;
+    gap: 10px;
+  }
+
+  &__button {
+    padding: 5px 10px;
+    background-color: #0000;
+    border: 1px solid black;
+    transition: 0.3s;
+    outline: none;
+    font-size: 15px;
+
+    &--active {
+      background-color: $hover-color;
+    }
+
+    &:hover {
+      background-color: $hover-color;
+    }
+  }
+}
