+++ src/styles/fonts.scss
@@ -1,0 +1,1 @@
+@import url("https://fonts.googleapis.com/css2?family=Secular+One&display=swap");
