+++ src/styles/variables.scss
@@ -1,0 +1,1 @@
+$hover-color: #ffcc70;
